requirements.txt
PyYAML == 3.12
ipython == 6.1.0
matplotlib == 1.5.3
numpy == 1.11.3
pandas == 0.18.1
seaborn == 0.7.1
boto3 == 1.4.6
scikit-learn == 0.19.0

-e git+https://github.com/felix-last/imbalanced-learn.git#egg=imbalanced-learn
-e git+https://github.com/felix-last/imbalanced-tools.git#egg=imbalanced-tools
